Tidy up PR.
