+++ src/providers/bytez/index.ts
@@ -2,62 +2,10 @@ import crypto from 'node:crypto';
 import { ParameterConfig, ProviderConfigs } from '../types';
 import BytezInferenceAPI from './api';
 import { BytezInferenceChatCompleteConfig } from './chatComplete';
+import { LRUCache } from './utils';
+import { BytezResponse } from './types';
 
 const BASE_URL = 'https://api.bytez.com/models/v2';
-
-class LRUCache<K, V> {
-  private size: number;
-  private map: Map<K, V>;
-
-  constructor({ size = 100 } = {}) {
-    this.size = size;
-    this.map = new Map();
-  }
-
-  get(key: K): V | undefined {
-    if (!this.map.has(key)) return undefined;
-
-    // Move the key to the end to mark it as recently used
-    const value = this.map.get(key)!;
-    this.map.delete(key);
-    this.map.set(key, value);
-    return value;
-  }
-
-  set(key: K, value: V): void {
-    if (this.map.has(key)) {
-      // Remove the old value to update position
-      this.map.delete(key);
-    } else if (this.map.size >= this.size) {
-      // Remove least recently used (first item in Map)
-      const lruKey: any = this.map.keys().next().value;
-      this.map.delete(lruKey);
-    }
-
-    // Insert the new key-value as most recently used
-    this.map.set(key, value);
-  }
-
-  has(key: K): boolean {
-    return this.map.has(key);
-  }
-
-  delete(key: K): boolean {
-    return this.map.delete(key);
-  }
-
-  keys(): IterableIterator<K> {
-    return this.map.keys();
-  }
-
-  values(): IterableIterator<V> {
-    return this.map.values();
-  }
-
-  get length(): number {
-    return this.map.size;
-  }
-}
 
 const IS_CHAT_MODEL_CACHE = new LRUCache({ size: 100 });
 
@@ -66,113 +14,87 @@ const BytezInferenceAPIConfig: ProviderConfigs = {
   chatComplete: BytezInferenceChatCompleteConfig,
   requestHandlers: {
     chatComplete: async ({ providerOptions, requestBody }) => {
-      const { model: modelId } = requestBody;
-
-      let adaptedBody;
-
       try {
-        adaptedBody = bodyAdapter(requestBody);
-      } catch (error: any) {
+        const { model: modelId } = requestBody;
+
+        const adaptedBody = bodyAdapter(requestBody);
+
+        const headers = {
+          'Content-Type': 'application/json',
+          Authorization: `Key ${providerOptions.apiKey}`,
+        };
+
+        const isChatModel = await validateModelIsChat(modelId, headers);
+
+        if (!isChatModel) {
+          return constructFailureResponse(
+            'Bytez only supports chat models on PortKey',
+            { status: 400 }
+          );
+        }
+
+        const url = `${BASE_URL}/${modelId}`;
+
+        const response = await fetch(url, {
+          method: 'POST',
+          headers,
+          body: JSON.stringify(adaptedBody),
+        });
+
+        if (adaptedBody.stream) {
+          return new Response(response.body, response);
+        }
+
+        const { error, output }: BytezResponse = await response.json();
+
+        if (error) {
+          return constructFailureResponse(error, response);
+        }
+
         return new Response(
           JSON.stringify({
-            status: 'failure',
-            message: error.message,
-          }),
-          {
-            status: 500,
-            headers: {
-              'content-type': 'application/json',
+            id: crypto.randomUUID(),
+            object: 'chat.completion',
+            created: Date.now(),
+            model: modelId,
+            choices: [
+              {
+                index: 0,
+                message: output,
+                logprobs: null,
+                finish_reason: 'stop',
+              },
+            ],
+            usage: {
+              inferenceTime: response.headers.get('inference-time'),
+              modelSize: response.headers.get('inference-meter'),
             },
-          }
-        );
-      }
-
-      const headers = {
-        'Content-Type': 'application/json',
-        Authorization: `Key ${providerOptions.apiKey}`,
-      };
-
-      const isChatModel = await validateModelIsChat(modelId, headers);
-
-      if (!isChatModel) {
-        return new Response(
-          JSON.stringify({
-            status: 'failure',
-            message: 'Bytez only supports chat models on PortKey',
-          }),
-          {
-            status: 500,
-            headers: {
-              'content-type': 'application/json',
-            },
-          }
-        );
-      }
-
-      const url = `${BASE_URL}/${modelId}`;
-
-      const response = await fetch(url, {
-        method: 'POST',
-        headers,
-        body: JSON.stringify(adaptedBody),
-      });
-
-      if (adaptedBody.stream) {
-        return new Response(response.body, response);
-      }
-
-      const { error, output }: { error: string | null; output: object | null } =
-        await response.json();
-
-      if (error) {
-        return new Response(
-          JSON.stringify({
-            //
-            message: error,
           }),
           response
         );
+      } catch (error: any) {
+        return constructFailureResponse(error.message);
       }
-
-      return new Response(
-        JSON.stringify({
-          id: crypto.randomUUID(),
-          object: 'chat.completion',
-          created: Date.now(),
-          model: modelId,
-          choices: [
-            {
-              index: 0,
-              message: output,
-              logprobs: null,
-              finish_reason: 'stop',
-            },
-          ],
-          usage: {
-            inferenceTime: response.headers.get('inference-time'),
-            modelSize: response.headers.get('inference-meter'),
-            // prompt_tokens: 11,
-            // completion_tokens: 28,
-            // total_tokens: 39,
-            // prompt_tokens_details: {
-            //   cached_tokens: 0,
-            //   audio_tokens: 0,
-            // },
-            // completion_tokens_details: {
-            //   reasoning_tokens: 0,
-            //   audio_tokens: 0,
-            //   accepted_prediction_tokens: 0,
-            //   rejected_prediction_tokens: 0,
-            // },
-          },
-          // service_tier: 'default',
-          // system_fingerprint: 'fp_34a54ae93c',
-        }),
-        response
-      );
     },
   },
 };
+
+function constructFailureResponse(message: string, response?: object) {
+  return new Response(
+    JSON.stringify({
+      status: 'failure',
+      message,
+    }),
+    {
+      status: 500,
+      headers: {
+        'content-type': 'application/json',
+      },
+      // override defaults if desired
+      ...response,
+    }
+  );
+}
 
 function bodyAdapter(requestBody: Record<string, any>) {
   for (const [param, paramConfig] of Object.entries(
@@ -242,16 +164,6 @@ function bodyAdapter(requestBody: Record<string, any>) {
   }
 
   return adaptedBody;
-}
-
-interface Model {
-  task: string;
-}
-
-interface BytezResponse {
-  error: string;
-  output: Model[];
-  // add other model properties as needed
 }
 
 async function validateModelIsChat(
