+++ src/providers/bytez/types.ts
@@ -1,0 +1,11 @@
+interface Model {
+  task: string;
+}
+
+interface BytezResponse {
+  error: string;
+  output: Model[];
+  // add other model properties as needed
+}
+
+export { Model, BytezResponse };
